Refactoring of ships array, generate btn added, simple peerjs example implemented
