+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\ammar.botonjic\\Documents\\battleship-project\\src\\index.js":"1","C:\\Users\\ammar.botonjic\\Documents\\battleship-project\\src\\App.jsx":"2","C:\\Users\\ammar.botonjic\\Documents\\battleship-project\\src\\renderCoordinates.helpers.js":"3","C:\\Users\\ammar.botonjic\\Documents\\battleship-project\\src\\constants\\constants.js":"4","C:\\Users\\ammar.botonjic\\Documents\\battleship-project\\src\\2DArray.styled.jsx":"5"},{"size":230,"mtime":1614278326201,"results":"6","hashOfConfig":"7"},{"size":2047,"mtime":1616326652171,"results":"8","hashOfConfig":"7"},{"size":2048,"mtime":1616184479270,"results":"9","hashOfConfig":"7"},{"size":23,"mtime":1616328129855,"results":"10","hashOfConfig":"7"},{"size":831,"mtime":1614278326195,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},"tzu57g",{"filePath":"15","messages":"16","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},"C:\\Users\\ammar.botonjic\\Documents\\battleship-project\\src\\index.js",[],["23","24"],"C:\\Users\\ammar.botonjic\\Documents\\battleship-project\\src\\App.jsx",["25"],"C:\\Users\\ammar.botonjic\\Documents\\battleship-project\\src\\renderCoordinates.helpers.js",[],"C:\\Users\\ammar.botonjic\\Documents\\battleship-project\\src\\constants\\constants.js",[],"C:\\Users\\ammar.botonjic\\Documents\\battleship-project\\src\\2DArray.styled.jsx",[],{"ruleId":"26","replacedBy":"27"},{"ruleId":"28","replacedBy":"29"},{"ruleId":"30","severity":1,"message":"31","line":52,"column":6,"nodeType":"32","endLine":52,"endColumn":8,"suggestions":"33"},"no-native-reassign",["34"],"no-negated-in-lhs",["35"],"react-hooks/exhaustive-deps","React Hook useEffect has a missing dependency: 'generateShips'. Either include it or remove the dependency array.","ArrayExpression",["36"],"no-global-assign","no-unsafe-negation",{"desc":"37","fix":"38"},"Update the dependencies array to be: [generateShips]",{"range":"39","text":"40"},[1094,1096],"[generateShips]"]
+[{"C:\\Users\\ammar.botonjic\\Documents\\battleship-project\\src\\index.js":"1","C:\\Users\\ammar.botonjic\\Documents\\battleship-project\\src\\App.jsx":"2","C:\\Users\\ammar.botonjic\\Documents\\battleship-project\\src\\renderCoordinates.helpers.js":"3","C:\\Users\\ammar.botonjic\\Documents\\battleship-project\\src\\constants\\constants.js":"4","C:\\Users\\ammar.botonjic\\Documents\\battleship-project\\src\\2DArray.styled.jsx":"5"},{"size":230,"mtime":1614278326201,"results":"6","hashOfConfig":"7"},{"size":1898,"mtime":1616418046260,"results":"8","hashOfConfig":"7"},{"size":2749,"mtime":1616415445467,"results":"9","hashOfConfig":"7"},{"size":23,"mtime":1616328129855,"results":"10","hashOfConfig":"7"},{"size":831,"mtime":1614278326195,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},"tzu57g",{"filePath":"15","messages":"16","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},"C:\\Users\\ammar.botonjic\\Documents\\battleship-project\\src\\index.js",[],["23","24"],"C:\\Users\\ammar.botonjic\\Documents\\battleship-project\\src\\App.jsx",["25"],"C:\\Users\\ammar.botonjic\\Documents\\battleship-project\\src\\renderCoordinates.helpers.js",[],"C:\\Users\\ammar.botonjic\\Documents\\battleship-project\\src\\constants\\constants.js",[],"C:\\Users\\ammar.botonjic\\Documents\\battleship-project\\src\\2DArray.styled.jsx",[],{"ruleId":"26","replacedBy":"27"},{"ruleId":"28","replacedBy":"29"},{"ruleId":"30","severity":1,"message":"31","line":31,"column":6,"nodeType":"32","endLine":31,"endColumn":8,"suggestions":"33"},"no-native-reassign",["34"],"no-negated-in-lhs",["35"],"react-hooks/exhaustive-deps","React Hook useEffect has a missing dependency: 'generateShips'. Either include it or remove the dependency array.","ArrayExpression",["36"],"no-global-assign","no-unsafe-negation",{"desc":"37","fix":"38"},"Update the dependencies array to be: [generateShips]",{"range":"39","text":"40"},[944,946],"[generateShips]"]
