+++ src/renderCoordinates.helpers.js
@@ -46,7 +46,46 @@ const isColision = (coordinate, shipArray) => {
   return result;
 }
 
-export const fillShipsCoordinates = (shipArray) => {
+export const fillShipsCoordinates = () => {
+  const shipArray = [
+    {
+      name: 'destroyer',
+      size: 2,
+      class: 'red',
+      coordinates: []
+    },
+    {
+      name: 'destroyer',
+      size: 2,
+      class: 'red',
+      coordinates: []
+    },
+    {
+      name: 'submarine',
+      size: 3,
+      class: 'purple',
+      coordinates: []
+    },
+    {
+      name: 'cruiser',
+      size: 3,
+      class: 'purple',
+      coordinates: []
+    },
+    {
+      name: 'battleship',
+      size: 4,
+      class: 'green',
+      coordinates: []
+    },
+    {
+      name: 'carrier',
+      size: 5,
+      class: 'yellow',
+      coordinates: []
+    },
+  ]
+
   shipArray.forEach((ship) => {
     let tempCoordinates = [];
     let doWork = true;
@@ -71,4 +110,5 @@ export const fillShipsCoordinates = (shipArray) => {
     }
     ship.coordinates.push(...tempCoordinates);
   })
+  return shipArray;
 }
